requirements.txt
numpy
scipy
matplotlib
h5py
qiskit-terra
qiskit-aer
bqskit
openfermion
jupyter
notebook